+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # Changelog
-## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [2.7rc0]
+## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [2.7rc1]
 This version of ONE adds support for Alyx 2.0.0 and pandas 3.0.0 with dataset QC filters. This version no longer supports 'data' search filter.
 
 ### Added
@@ -15,7 +15,7 @@ This version of ONE adds support for Alyx 2.0.0 and pandas 3.0.0 with dataset QC
 - support 'qc' category field in dataset cache table
 - One.search supports ´dataset_qc_lte` filter
 - One.list_datasets supports ´dataset_qc_lte` and `ignore_qc_not_set` filters
-- one.alf.io.iter_sessions more performant
+- one.alf.io.iter_sessions pattern arg to make more performant
 
 ### Removed